[PATCH] md: Always allow a half-built md array to be stopped.
When starting an array fails, we have to tear it down, but
in some circumstances (particularly autostart_array) the
reference count will be 3, so do_md_stop will fail.

With this patch we only worry about the number of users
is the array has been fully started.

BKrev: 3ed2e3f1bbNWOya-tCLpJ0DaVRBInw