Merge bk://kernel.bkbits.net/jgarzik/net-drivers-2.5 into home.transmeta.com:/home/torvalds/v2.5/linux
2003/05/27 01:14:11-04:00 jgarzik
[netdrvr eepro] update MODULE_AUTHOR per old-author request

2003/05/27 01:05:58-04:00 edward_peng
[netdrvr sundance] fix another flow control bug

2003/05/27 01:04:28-04:00 edward_peng
[netdrvr sundance] fix flow control bug

2003/05/27 01:01:53-04:00 bunk
[wan lmc] remove 2.0.x-era code

The patch below removes obsolete #if'd code for kernel 2.0 and 2.2 from
drivers/net/wan/lmc/* (this includes the expansion of some #define's
that were definded differently for different kernel versions).

2003/05/27 00:45:17-04:00 shmulik.hen
[netdrvr bonding] fix ABI version control problem

This fix makes bonding not commit to a specific ABI version if the ioctl
command is not supported by bonding.

(It also removes the '\n' in the continuous printk reporting the link down
event in bond_mii_monitor - it got in there by mistake in our previous
patch set and caused log messages to appear funny in some situations).

2003/05/27 00:42:05-04:00 shmulik.hen
[netdrvr bonding] fix long failover in 802.3ad mode

This patch fixes the bug reported by Jay on April 3rd regarding long
failover time when releasing the last slave in the active aggregator. The
fix, as suggested by Jay, is to follow the spec recommendation and send a
LACPDU to the partner saying this port is no longer aggregatable and
therefore trigger an immediate re-selection of a new aggregator instead of
waiting the entire expiration timeout.

BKrev: 3ed2fef6UtwCMt0vCiarMkrUfND3DA