Merge bk://kernel.bkbits.net/acme/usb-2.5 into kroah.com:/home/greg/linux/BK/gregkh-2.5
2003/05/29 15:25:37-07:00 hch
[PATCH] fix scsi_register_host abuse in usb scanner drivers

They should be using scsi_add_host directly.  I had to rewrite
half of the drivers, though to fix horrible braindamage like
leaving dangling scsi structures around after ->disconnect.

Gettig rid of the remaining scsi_register_host callers is required
for the scsi stack to move forward so please try to forward this
to Linus in a timely mannor, thanks!

2003/05/29 15:25:26-07:00 hch
[PATCH] use second arg to scsi_add_host in usb storage

That way we don't need the addition scsi_set_device call.

2003/05/29 15:19:51-07:00 mdharm-usb
[PATCH] USB: storage: collapse one-use functions

This patch collapses some one-use functions into their callers.  It also
clones some code for control transfers so we can implement abortable
control transfers with timeout.

This patch is from Alan Stern.

 Remove usb_stor_bulk_msg() and usb_stor_interrupt_msg().  Move their
 functionality into usb_stor_bulk_transfer_buf() and
 usb_stor_intr_transfer().

 Move the functionality of usb_stor_control_msg() into
 usb_stor_ctrl_transfer().

 Remove the unused act_len parameter from usb_stor_intr_transfer().

2003/05/29 15:19:41-07:00 mdharm-usb
[PATCH] USB: storage: abort and disconnect handling.

This patch re-organizes abort handling and enhances disconnect handling.

Not only do we keep track of the state (ABORTING, IDLE, etc.), but during
an abort we now introduce the idea of 'okay to send' or not.  The idea is
that we can now implement reset-after-abort properly.

We also track if we're disconnecting, and use that data to determine if we
can submit URBs or not.  Which means we can now disconnect during an abort.

This patch is from Alan Stern.

2003/05/29 15:17:04-07:00 oliver
[PATCH] USB: allocate memory for reset earlier

if we fail with -ENOMEM, we should do it before the device must be
reparsed.

2003/05/29 15:03:07-07:00 greg
USB: fix up unusual_devs.h merge mess

Thanks to Per Winkvist for the info and patches to do this.

2003/05/29 14:53:47-07:00 greg
USB: build gadget drivers if they are built into the kernel.

Somehow this got merged away...

BKrev: 3ed6e360LVkxXO6p4x_ctadIE_6qAA