[PATCH] Fix disk partitioning with multiple IDE disks
Now that we use idedefault_driver.drives instead of ata_unused, the order
of drives on the driver->drives becomes significant; note that when we added
to ata_unused, we had done that to the end of list.

So we should do the same with the "drives" list.

BKrev: 3edab264Z2oRQV5yfIMKpNnzkHLVtQ