Merge bk://kernel.bkbits.net/acme/net-2.5 into nuts.ninka.net:/home/davem/src/BK/net-2.5
2003/05/30 23:55:21-03:00 acme
o n_hdlc: CodingStyle cleanups and removal of old stuff

- remove include kerneld.h, long gone
- some 80 column reformatting
- removal of typedefs
- c99 style documentation, using the existing function documentation
- add MODULE_AUTHOR
- remove GET_USER, COPY_TO_USER and friends
- remove ssize_t typedef, already in the kernel headers
- remove rw_count_t and rw_ret_t, use simple int and size_t as in the
  tty ops prototypes and avoiding casts for return
- make strings in printk in init/exit module routines __initdata/__exitdata
  saving some bytes after module init and for the case where this driver is
  statically linked in the kernel (__exitdata is trown away)

This makes this driver comply with parts of CodingStyle (more to be done) and
makes it looks more like the rest of the kernel, making it easier to read/debug.

Have been using this modifications with ADSL/syncppp for some time, working well.

BKrev: 3edd74faDStk2uLDupnBYohgQmqTig