[PATCH] allow thread_info to be allocated as part of task_struct
This re-organizes the thread info allocations a bit, so that
architectures like ia64 that want to allocate the thread_info structure
as part of the task structure allocation can do so.

The bulk of the patch is adding the "tsk" information to the thread
info allocator (ignored by all non-ia64 architectures), and switching
the order of the allocators to make this all possible.

BKrev: 3eded6a8UbXse96M4jfzl7Iqnxga0g