Merge bk://ldm.bkbits.net/linux-2.5-core into home.transmeta.com:/home/torvalds/v2.5/linux
2003/06/05 15:16:20-07:00 mochel
[driver model] Remove extraneous class device release method.

What happens when you get a patch that does something an applied patch
already does, but a little better, and merge it sloppily: You end up calling
the wrong function because you've defined equivalent methods in two places.

Bad Pat, Bad.

2003/06/05 14:48:21-07:00 mochel
Merge osdl.org:/home/mochel/src/kernel/devel/linux-2.5-virgin
into osdl.org:/home/mochel/src/kernel/devel/linux-2.5-core

2003/06/05 13:48:47-07:00 mochel
[kobject] Add warning + back trace if kobject_get() is called with 0 refcount.

2003/06/05 13:25:00-07:00 mochel
[driver model] Add device_for_each_child iterator.

From Mike Anderson:

I have been using it on an outstanding patch for scsi_set_host_offline. It
appears to work fine in my testing.

2003/06/05 09:23:44-07:00 mochel
[kobject] Remove kobj_lock and use lockless refcounting.

The only thing preventing this from happening earlier was the circular sysfs
registration dependency - it would need to be initialized before it was
registered, but needed to be registered before it was initialized.

With kobjects gone from struct filesystem_type, the dependency no longer
exists and we don't have to special-case the possibility that a kobject will
be passed to kobject_get with a refcount == 0.

Note that a kobject with a count of 0 in that function is still a bug, but
one in the subsystem making the call. We should add a debugging hook to dump
the stack if it does happen.

2003/06/05 08:55:13-07:00 mochel
[fs] Remove kobject support for filesystems

It was initially added for the immediate gain of being able to see what
filesystems were registered in the system. However, it was done without
proper analysis of the lifetime rules associated with them.

Someday, we should convert struct filesystem_type and struct super_block to
use kobjects, though it will take sufficent time and effort to get it right.
Until then, we go without..

BKrev: 3edfc3c5JJCf4mVsclEZsWWwFwXwgA