Merge redhat.com:/garz/repo/linus-2.5 into redhat.com:/garz/repo/net-drivers-2.5
2003/06/06 01:13:35-04:00 yoshfuji
[netdrvr] C99 initializers for arcnet

2003/06/06 01:08:11-04:00 scott.feldman
[PATCH] remove ethtool privileged references

dev_ioctl already checks capable(CAP_NET_ADMIN) for SOICETHTOOL, so
privileged reference are not necessary.

2003/06/06 01:07:46-04:00 scott.feldman
[PATCH] 10GbE ethtool support

Add 10GbE support for ethtool.

2003/06/06 01:07:21-04:00 zwane
[PATCH] cli/sti cleanup for fmvj18x

This one should be safe as we're protected by the xmit_lock in all instances

2003/06/06 01:04:47-04:00 jgarzik
[netdrvr] add MAINTAINERS entry for atmel wireless driver

2003/06/06 01:03:23-04:00 srk
[netdrvr] add atmel[_cs], new wireless driver

Attached is a driver for Atmel at76c50x WiFi cards. This code started
out as a GPL release from Atmel of pretty horrible quality and I've
extensively re-worked it with the aim of making it acceptable in the
kernel. Please could you take a look and either pass it into the patch
stream or let me know what's wrong with it?

The code has been tested on at least three different brand cards by
different people. Jean Tourrilhes took a look at an earlier version an
was positive. He's put incorporating this into 2.6 as a priority 1.
The patch works fine on 2.5.70.

The firmware issue has been addressed now. The only firmware in the
driver is a small stub which reads the MAC address from NVRAM on the
card. The source for that is included so there are no GPL issues. The
main firmware is loaded from userspace using Manuel Estrada Sainz's
sysfs firmware class. I know that the  patch for that has been
accepted but it hasn't turned up anywhere I can see yet. The
driver compiles fine even without the firmware class. I've made a
package of the firmware images which is available from my website.

The remaining issues with the driver are migrating PCMCIA to the new
driver model and PCI support. I'm happy to produce followup patches as
the PCMCIA system gets evolved to the new driver model: the timing on
that is controlled by others. This set of chips includes a PCI version
and the driver should support that, but AFAIK there is no PCI hardware
available anywhere. If Atmel can provide me with some it will be
simple to add PCI support.

The driver uses the CRC32 library module and the firmware loader. I've
not put in dependencies on those, but when the lastest set of patches
go into Kconfig I'll set it up so that selecting the Atmel driver
selects CRC32 and FW_LOADER too.

2003/06/06 00:52:32-04:00 shemminger
[PATCH] sb1000 driver bugs

Inspecting the sb1000 driver showed some interesting bugs:
	- net device pointer is used before the device is allocated; gcc
	  does catch this.
	- unregister is called even though device not registered successfully
	- net device is not freed on remove.

Compiles but don't have hardware to test.  Don't know how it ever worked though.

2003/06/05 00:38:08-04:00 reeja.john
[netdrvr amd8111e] link against mii lib

2003/06/05 00:36:38-04:00 jgarzik
[netdrvr skge] add ULL modifier to 64-bit constant

2003/06/05 00:01:42-04:00 jgarzik
[netdrvr] gcc 3.3 cleanups

Mostly adding 'ULL' modifier to 64-bit constants.

BKrev: 3ee0268dz5PgI_nxcFjcT6oVfEHCMg