Merge kroah.com:/home/greg/linux/BK/bleed-2.5 into kroah.com:/home/greg/linux/BK/gregkh-2.5
2003/06/10 14:50:54-07:00 david-b
[PATCH] USB: ohci-hcd, remove FIXME

The endpoint disable() change resolved the bug
identified by the FIXME ... this updates the comment.

2003/06/10 14:50:17-07:00 david-b
[PATCH] USB: net2280 patch: control-out fix, minor cleanups

The main thing this fixes is making the control-OUT path work.
Drivers like RNDIS and DFU need it; this should resolve one
bug report.  It also has some minor cleanups.

2003/06/10 14:49:50-07:00 david-b
[PATCH] USB: usb/core/devio: identify process

Rather than just saying that USBDEVFS_CONTROL failed,
say also which process it failed for ... so it's easier
to figure out what's happen.

2003/06/10 14:49:35-07:00 stern
[PATCH] USB: Make hub.c DMA-aware

This patch makes the hub status irq DMA-aware, by pre-allocating the
transfer buffer in consistent memory.  Unfortunately, there doesn't seem
to be an easy way to do the same for the status report buffers.

2003/06/10 14:36:33-07:00 mdharm-usb
[PATCH] USB: usb-storage: remove dead code

This patch removes the code that faked the EVPD INQUIRY.  The SCSI core no
longer does that, so there is no need to filter it.

2003/06/10 14:35:56-07:00 mdharm-usb
[PATCH] USB: usb-storage: re-organize probe/disconnect

This patch re-organizes probe and disconnect into smaller functions (which
are all functionally equivalent to the current code).  This allows easier
verification that the code is correct, and will make for easier
implementation of the proper SCSI shutdown code.

2003/06/10 14:35:30-07:00 mdharm-usb
[PATCH] USB: usb-storage: handle babble

This patch introduces some handling for babble conditions.

Basically, once a babble is detected, we return sense data saying the
command was invalid.  We also go on to transfer the CSW (for BBB transport)
so we stay in phase with the device.

This isn't guaranteed to work with every device that babbles, but it can't
hurt compared to the current behavior.  Properly operating devices are
unaffected by this patch.

2003/06/10 14:35:17-07:00 olh
[PATCH] USB: incorrect ethtool -i driver name

2003/06/06 15:08:08-07:00 petkan
[PATCH] USB: pegasus patch

  better error handling and ethtool ioctl() cleanup.  HOME_PNA
  now should work (at least for the pegasus II based devices).

  one more vendor and device IDs.

2003/06/06 15:03:24-07:00 stern
[PATCH] USB: Rename static functions in hub.c and increase timeouts

As requested by David Brownell, this patch removes the usb_ prefix from
the static functions in hub.c.  It also multiplies the timeouts for
GET_STATUS and GET_DESCRIPTOR control transfers by USB_CTRL_GET_TIMEOUT.

2003/06/06 15:03:07-07:00 stern
[PATCH] USB: Don't allocate transfer buffers on the stack in hub.c

 Allocate hub status input buffers separately and not on the stack.

2003/06/06 15:02:55-07:00 oliver
[PATCH] USB: kill a compiler warning in hpusbscsi

  - kill useless variable

2003/06/06 15:02:42-07:00 joe
[PATCH] USB: vicam.c patch

I noticed a version of vicam.c a few revisions ago had all the /proc fs
writing removed because I was incorrectly using potentially tainted user
space pointers.  Here's a patch which I think fixes the pointer issue and
restores the /proc fs interface to vicam.  If this fix is still
problematic, please let me know and I'll fix whatever comes up.

BKrev: 3ee65a15HpqqV5g7awO29Pg11dr_Kg