Merge bk://kernel.bkbits.net/gregkh/linux/linus-2.5 into home.transmeta.com:/home/torvalds/v2.5/linux
2003/06/10 16:06:00-07:00 torvalds
Don't make the source checker default path be quite
so hackish.

2003/06/10 15:46:36-07:00 torvalds
Merge bk://ldm.bkbits.net/linux-2.5-core
into home.transmeta.com:/home/torvalds/v2.5/linux

2003/06/10 15:27:47-07:00 mochel
hand merge

2003/06/10 15:25:48-07:00 mochel
Hand merge

2003/06/10 14:53:51-07:00 mochel
[sysfs] Get zeroed page for file read/write buffers.

2003/06/10 14:32:08-07:00 mochel
[cpufreq] Convert to use new system device API

- Remove explicit call from arm PM sequence, as its handled implicitly
  by sysdev_restore() in driver model core.

2003/06/10 14:28:54-07:00 mochel
[mtrr] Add save()/restore() methods.

Patch originally from Nigel Cunningham and Pavel Machek. Cleaned up and
converted to new system device API by your truly.

2003/06/10 14:27:26-07:00 mochel
[cpu] Use sysdev.h instead of device.h and export cpu_sysdev_class

2003/06/10 14:26:43-07:00 mochel
[driver model] Don't Oops when registering global sysdev drivers.

2003/06/10 13:58:37-07:00 mochel
[driver model] Add save() and restore() methods for system device drivers.

It turns out that at least some system device drivers need to allocate
memory and/or sleep for one reason or another when either saving or
restoring state.

Instead of adding a 'level' paramter to the suspend() and resume() methods,
which I despise and think is a horrible programming interface, two new
methods have been added to struct sysdev_driver:

        int     (*save)(struct sys_device *, u32 state);
        int     (*restore)(struct sys_device *);

that are called explicitly before and after suspend() and resume()
respectively, with interrupts enabled. This gives the drivers the
flexibility to allocate memory and sleep, if necessary.

2003/06/10 12:10:39-07:00 mochel
[s390 xpram] Use sysdev.h instead of device.h

2003/06/10 12:08:04-07:00 mochel
[x86-64 i8259] Use sysdev.h instead of device.h

2003/06/10 12:06:44-07:00 mochel
[oprofile] Use sysdev.h instead of device.h

2003/06/10 12:05:03-07:00 mochel
[timer] Use sysdev.h instead of device.h

2003/06/10 12:04:07-07:00 mochel
[nmi] Use sysdev.h instead of device.h

2003/06/10 12:03:25-07:00 mochel
[i8259] Use sysdev.h instead of device.h

2003/06/10 12:02:44-07:00 mochel
[apic] Use sysdev.h instead of device.h

2003/06/10 12:01:36-07:00 mochel
[driver model] Remove system device definitions from device.h

Should have been in earlier changeset. D'oh.

2003/06/10 12:00:51-07:00 mochel
[numa nodes] Convert to use new system device API

2003/06/10 12:00:07-07:00 mochel
[memblk] Convert to use new system device API

2003/06/10 11:59:13-07:00 mochel
[driver model] Make sure right header is used for cpu.c

2003/06/10 11:58:22-07:00 mochel
[driver model] Create include/linux/sysdev.h and define sysdev_attribute.

Split out all system device definitions from device.h into their own header
sysdev.h

Define struct sysdev_attribute and define functions to export attributes
in sysfs.

2003/06/10 09:45:45-07:00 mochel
[s390 xpram] Convert to use new system device API.

2003/06/10 09:44:23-07:00 mochel
[x86-64 i8259] Convert to use new system device API.

2003/06/10 09:43:42-07:00 mochel
[oprofile] Convert to use new system device API.

2003/06/10 09:43:00-07:00 mochel
[timer] Convert to use new system device API.

2003/06/10 09:42:22-07:00 mochel
[nmi] Convert to use new system device API.

2003/06/10 09:41:55-07:00 mochel
[i8259] Convert to use new system device API.

2003/06/10 09:36:33-07:00 mochel
[lapic] Convert to new system device API.

2003/06/09 16:44:27-07:00 mochel
[driver model] Convert to new system device API

2003/06/09 16:43:08-07:00 mochel
[driver model] Make sure that system devices are handled specially power-wise

- Suspend system devices last, after interrupts have been enabled.
- Resume them first, before interrupts are enabled.
- Shut them down last, after everything else.

2003/06/09 16:41:22-07:00 mochel
[kobject] Add set_kset_name

Shorthand macro for initializing only the name of an embedded kset in an
object.

2003/06/09 16:40:34-07:00 mochel
[list.h] Add list_for_each_entry_reverse

2003/06/09 16:40:02-07:00 mochel
[driver model] Rewrite system device API

System devices are special, and after two years of listening to Linus
preach this, it finally sunk in enough to do something about. We don't
need to regard them as real devices that reside on a peripheral bus and
can be dynamically bound to drivers. If we discover, e.g. a CPU, we know
by default that we have a driver for it, and we know damn well that we
have a CPU. We still need to keep track of all the devices, and all the
devices of a particular type. The kobject infrastructure allows us to do
this, without the overhead of the regular model.

A new subsystem is defined that registers as a child object of
devices_subsys, giving us:

        /sys/devices/system/

struct sysdev_class {
        struct list_head        drivers;

        /* Default operations for these types of devices */
        int     (*shutdown)(struct sys_device *);
        int     (*suspend)(struct sys_device *, u32 state);
        int     (*resume)(struct sys_device *);
        struct kset             kset;
};

Defines a type of system device. These are registered on startup, by e.g.
drivers/base/cpu.c. The methods are default operations for devices of that
type that may or may not be used. For things like the i8259 controller,
these will be filled in, since it is registered by the same component that
the device controls reside in.

For things like CPUs, generic code will register the class, but other
architecture-specific or otherwise configurable drivers may register
auxillary drivers, that look like:

struct sysdev_driver {
        struct list_head        entry;
        int     (*add)(struct sys_device *);
        int     (*remove)(struct sys_device *);
        int     (*shutdown)(struct sys_device *);
        int     (*suspend)(struct sys_device *, u32 state);
        int     (*resume)(struct sys_device *);
};


Each auxillary driver gets called during each operation on a device of a
particular class.
Auxillary drivers may register with a NULL class parameter, in which case
they will be added to a list of 'global drivers' that get called for each
device of each class.


Besides providing a decent of cleanup for system device drivers, this also
allows:

- Special handling of system devices during power transitions.

  We no longer have to worry about shutting down the PIC before we shut
  down any devices. We can shut down the system devices after we've shut
  down every other device.

  Ditto for suspend/resume cycles. Almost (if not) all PM actions for
  system devices happen with interrupts off, and require only one call,
  which makes that easier. But, we can also make sure we take care of
  these last during suspend and first during resume.

- Easy expression of configurable device-specific interfaces.

  Namely cpufreq and mtrr. We don't have to worry about mispresentation in
  the driver model (like recent MTRR patches) or using a cumbersome
  interface ({device,class}_interface) that don't receive all the
  necessary calls.

- Consolidation of userspace representation.

  No longer do we have /sys/devices/sys, /sys/bus/sys, and /sys/class/cpu,
  etc. We have only /sys/devices/system:

# tree /sys/devices/system/
/sys/devices/system/
|-- cpu
|   `-- cpu0
|-- i8259
|   `-- i82590
|-- lapic
|   `-- lapic0
|-- rtc
|   `-- rtc0
`-- timer
    `-- timer0

Each directory in 'system' is the class, and each directory under that is
the instance of each device in that class.

BKrev: 3ee67af7Eau-XmCS4l3Wd1bVnrbsrw