[PATCH] OProfile: Export task->tgid in the buffer
Export the task->tgid to userspace as well. This is needed
for forthcoming thread profiling stuff and should have been
done in the original patch ... oh well.

This requires an upgrade to oprofile 0.5.3. You can get it from
the website, or, for the impatient, here :

	http://movementarian.org/oprofile-0.5.3.tar.gz

BKrev: 3ee7f33bnfJt8ruWWDPq466q1eOOIg