Merge bk://kernel.bkbits.net/davem/net-2.5 into home.transmeta.com:/home/torvalds/v2.5/linux
2003/06/15 09:32:35-07:00 trond.myklebust
[PATCH] Fix rpc dentry list usage

This fixes the problems with NFS that got triggered by the list
poisoning.

The problem is that whereas using the dentry->d_hash list for private
purposes was previously harmless, it is not allowed with the RCU lists,
because we depend on the unhashed node continuing to point back to the
original list.

So keep a small vector of dentries around instead of using the d_hash
list to keep track of them.

BKrev: 3ef48540gmUlh6ZYM2zL3_ybzTYLow