Merge bk://kernel.bkbits.net/acme/net-2.5 into nuts.ninka.net:/home/davem/src/BK/net-2.5
2003/06/17 13:40:22-03:00 acme
o sock: remove sk_prev

Move it to the protocols that we're using this pointers for other
purposes than a list pointer as the name implies, namely tcp and
sctp, where they are used as a pointer to the bind_hash. Shrink,
struct sock, shrink! :-)

BKrev: 3ef485676DYm8A6KCV615jWxhbjJxw