[PATCH] kNFSd: Allow nfsv4 readdir to return filehandle when a mountpoint is found is a directory
From: "William A.(Andy) Adamson" <andros@citi.umich.edu>

When readdir is enumerating a directory and finds a mountpoint,
it needs to do a bit of extra work to find the filehandle to be
returned in the readdir reply.

It is even possible that finding the filehandle requires an up-call,
so the request might be dropped to be re-tried later.

BKrev: 3ef485760smrxhv5JOxPlaXbRrSMtA