[PATCH] JBD: remove jh_splice_lock
This was a strange spinlock which was designed to prevent another CPU from
ripping a buffer's journal_head away while this CPU was inspecting its state.

Really, we don't need it - we can inspect that state directly from bh->b_state.

So kill it off, along with a few things which used it which are themselves
not actually used any more.

BKrev: 3ef4858cxKIpK8b8alV2R_09yCoj2Q