[PATCH] JBD: remove journal_datalist_lock
This was a system-wide spinlock.

Simple transformation: make it a filesystem-wide spinlock, in the JBD
journal.

That's a bit lame, and later it might be nice to make it per-transaction_t.
But there are interesting ranking and ordering problems with that, especially
around __journal_refile_buffer().

BKrev: 3ef485942iWXXqolEYrLQtB2PDb69g