[PATCH] JBD: implement j_barrier_count locking
We now start to move onto the fields of the topmost JBD data structure: the
journal.

The patch implements the designed locking around the j_barrier_count member.
And as a part of that, a lot of the new locking scheme is implemented.
Several lock_kernel()s and sleep_on()s go away.

BKrev: 3ef48599-DEAzjoZReJ8p7yKYmdxsQ